Extract Team from TeamList.
This separation is necessary in order to be able to test the Team
component.
